(LOG-11874) - Removido antigo datepicker
