+++ src/components/inputs/index.js
@@ -4,7 +4,6 @@ import LInputSolo from './LInputSolo'
 import LSelectSolo from './LSelectSolo'
 import LDatePickerMonth from './LDatePickerMonth'
 import LDatePickerDay from './LDatePickerDay'
-import LDatePicker from './LDatePicker'
 import LDropdownCheck from './LDropdownCheck'
 import LSelectNamed from './LSelectNamed'
 import LInputLoaded from './LInputLoaded'
@@ -21,7 +20,6 @@ export {
     LSelectSolo,
     LDatePickerMonth,
     LDatePickerDay,
-    LDatePicker,
     LDropdownCheck,
     LSelectNamed,
     LInputLoaded,
